Move Webpack & Rollup configs back to root folder
