+++ rollup.config.js
@@ -1,5 +1,5 @@
 import babel from 'rollup-plugin-babel';
-import pkg from '../package.json';
+import pkg from './package.json';
 
 export default [
     {
