--- webpack/webpack.config.dev.js
@@ -1,5 +1,0 @@
-const base = require('./webpack.config.base');
-
-const config = base('development');
-
-module.exports = config;
