--- webpack/webpack.config.prod.js
@@ -1,5 +1,0 @@
-const base = require('./webpack.config.base');
-
-const config = base('production');
-
-module.exports = config;
